proposal: a niave rpc over tcp
So the current state of this program is heavily dependent on generics.
What the initial goal was to define a public struct Raft<T> that can act
as a consensus layer coupled along with any other type defined in rust.
With generics there obviously comes the issue of deserialization of some
requests, namely `AppendRequestsRPC` which communicates the log diffs
across nodes.

Initial implementation with docs.rs/tarpc in my old repo bwaklog/bee was
what I planned to use here, but u cant use generics.

What I plan to do, for literal fun, is to imlement some sort of rpc
calls using rust enums capturing different struct types. All request are
encoded to bytes with docs.rs/bincode and sent over TCP

Plans to call a general method Client::send taking in the RequestPattern
along with a client addr which returns the serialized response. This
pattern could be a bit better if I probably implement a service trait
for the client but this works for now

---

(17/12/24) NOTE: A lot of undeterministic behaviour as it stands. Append
Entries only sends empty entries to all nodes, just for the sake of testing
for now. The cluster remains stable up till a point, and after that all
nodes start fighting over leadership like children. Need to read the TLA
spec for raft and the paper carefully. Good progress so far

---

**changes
- (wip) moving to tokio async runtime
- implement PingRPC (some unwraps not handled)
- raft state persistence
- RequestVote RPC (WIP and note tested)
- AppendEntries RPC (WIP)

---

0430 hrs me says hi :)