feat: Log Replication (testing remaining)
Nodes are finally able to replicate log entries across a cluster. Have
not fully tested the algorithm, but 1% of it works in which the first
leader in the cluster is able to replicate entires.

No leader comitting logic has been written, this is just to save my
progress so far. Have initially made the mistake of using 0 as the base
index of my logs and not 1. Well in rust you index an array from 0, but
all the comparisions are made where the theoretical index is an offset
from the original index.